services/ai-processor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.2
python-dotenv==1.0.0
transformers==4.36.0
torch==2.1.1
scikit-learn==1.3.2
numpy==1.26.2
redis==5.0.1
httpx==0.25.2